Operator pods fix (#16)
* optimized test script

* Updated the execution type to use a phase parameter similar to pods
Also modified the execution logic to soldier on, even if too many pods
are created. This is is an intentional bug and is a holdover until a new
solution is completed. It means that pipelines can complete as long as
no created pods are in an error state.

* using fmt.Sprintf and filepath.Join to remove string arithmetic. Status still not updating

* using constants instead of string literals

* moving from pods to jobs because of added benefits from the job controller

* moved away from pods completely in favour of jobs.

* Migrated operator behaviour fully over to jobs
Configured skaffold for operator, and working through apparent RBAC,
issue regarding PVs.

* merge branch master

* moved UI away from pods, now rendering based on jobs

* Added more comprehensive task view too UI
UI now has clickable tasks, which can display the name, image, command,
and information about the duration. Also did some housekeeping, around
optimizing docker build processes

* fixed bug with execution status not being updated

* Begun work on logging functionality, fleshed out UI.
Refactored repeating logic to do with generating status icons for
pipelines. Implemented execution status in the pipeline list to show
overall status of executions. Pipelines and Executions may need to be polled
periodically to give a feeling of watching the system as it works.

* added colours for ui spinners

* refactored out a lot of functionality from main controller loop

* fixed pvc allocation, LGTM

* removed logs? message